+++ src/localconf.js
@@ -1,8 +1,6 @@
 /* eslint-disable prefer-destructuring */
 const { components, requiredCerts, config } = require('./params');
 
-const semver = require('semver');
-
 /**
  * Setup configuration variables based on the "app" the data-portal is
  * being deployed into (Brain Health Commons, Blood Pack, ...)
@@ -90,30 +88,8 @@ function buildConfig(opts) {
   const jobAPIPath = `${hostname}job/`;
   const credentialCdisPath = `${userAPIPath}credentials/cdis/`;
 
-  const coreMetadataPathPromise = fetch(peregrineVersionPath)
-    .then(response => response.json())
-    .then(responseBody => {
-      // if peregrine is on version 3.2.0/2023.04 or newer, or on a branch, use
-      // the peregrine endpoint. if not, use the deprecated pidgin endpoint
-      const minSemVer = '3.2.0';
-      const minMonthlyRelease = semver.coerce('2023.04.0', { loose: true });
-      const monthlyReleaseCutoff = semver.coerce('2020', { loose: true });
-
-      var url = `${hostname}api/search/coremetadata/`;
-      try {
-        var peregrineVersion = responseBody.version;
-        if (peregrineVersion) {
-          peregrineVersion = semver.coerce(peregrineVersion, { loose: true });
-          if (
-            semver.lt(peregrineVersion, minSemVer) ||
-            (semver.gte(peregrineVersion, monthlyReleaseCutoff) && semver.lt(peregrineVersion, minMonthlyRelease))
-          ) {
-            url = `${hostname}coremetadata/`; // pidgin endpoint
-          }
-        }
-      } catch (error) { } // can't parse or compare the peregrine version: don't use legacy url
-      return url;
-    });
+  const coreMetadataPath = `${hostname}api/search/coremetadata/`;
+  const coreMetadataLegacyPath = `${hostname}coremetadata/`;
 
   const indexdPath = typeof indexdURL === 'undefined' ? `${hostname}index/` : ensureTrailingSlash(indexdURL);
 
@@ -525,11 +501,13 @@ function buildConfig(opts) {
     apiPath,
     submissionApiPath,
     credentialCdisPath,
-    coreMetadataPathPromise,
+    coreMetadataPath,
+    coreMetadataLegacyPath,
     indexdPath,
     cohortMiddlewarePath,
     gwasWorkflowPath,
     graphqlPath,
+    peregrineVersionPath,
     dataDictionaryTemplatePath,
     graphqlSchemaUrl,
     appname: components.appName,
